Changes ready to test on CF
This requires `jq` to parse the JSON in the env vars
so there's no getting around building our own image.

This  uses .docker/foundryish.sh to run a script that munges
nuxeo.conf to the correct values appended AND comments out the first
use of those configuration
keys since it seems first use wins.

This should be change use a new startup
that does set env vars so the credentials aren't
in plain text